fix for 64-bit arch: use %zd to print size_t types and sizeof()
Approved by:	alfred
Tested on:	sparc64, amd64


Former-commit-id: a6b909e2cbc7b1ff6986ed7dce079781a3fef17f

+++ sys/nfs4client/nfs4_dev.h
@@ -95,7 +95,7 @@ begin_define
 define|#
 directive|define
 name|NFS4DEV_VERSION
-value|(0x3<< 16 | sizeof(struct nfs4dev_msg))
+value|(0x3<< 16 | (int) sizeof(struct nfs4dev_msg))
 end_define
 
 begin_comment
